+++ src/test/java/seedu/address/logic/parser/ClientNoteAddCommandParserTest.java
@@ -3,6 +3,7 @@ package seedu.address.logic.parser;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.fail;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NOTE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
@@ -24,17 +25,44 @@ class ClientNoteAddCommandParserTest {
     private static final String SPACE = " ";
     private static final String INDEX_STRING = "1";
     private static final String NOTE_STRING = "yay this be a note";
+    private static final String NO_INDEX_NO_NOTE = "just an empty string here";
+    private static final String HAS_INDEX_NO_NOTE = INDEX_STRING + " nope no note here";
+    private static final String HAS_INDEX_HAS_NOTE = INDEX_STRING + SPACE + PREFIX_NOTE + NOTE_STRING;
 
     private final TagNoteMap tagNoteMap = new TagNoteMap();
+    private final ClientNoteAddCommandParser parser = new ClientNoteAddCommandParser(tagNoteMap);
     private final Tag testTag = new Tag("testTag");
     private final Set<Tag> untaggedTags;
     private final Set<Tag> tags;
 
-    {
+    public ClientNoteAddCommandParserTest() {
         tags = new HashSet<>();
         tags.add(testTag);
         untaggedTags = new HashSet<>();
         untaggedTags.add(Tag.UNTAGGED);
+    }
+
+    @Test
+    public void parse_noIndexNoNote_throwsParseException() {
+        assertThrows(ParseException.class, () -> parser.parse(NO_INDEX_NO_NOTE));
+    }
+
+    @Test
+    public void parse_hasIndexNoNote_throwsParseException() {
+        assertThrows(ParseException.class, () -> parser.parse(HAS_INDEX_NO_NOTE));
+    }
+
+    @Test
+    public void parse_hasIndexHasNote_equalsExpected() {
+        try {
+            Note expectedClientNote = new Note(NOTE_STRING);
+            expectedClientNote.setTags(untaggedTags);
+            ClientNoteAddCommand expected = new ClientNoteAddCommand(Index.fromOneBased(1), expectedClientNote);
+            ClientNoteAddCommand actual = parser.parse(HAS_INDEX_HAS_NOTE);
+            assertEquals(expected, actual);
+        } catch (ParseException e) {
+            fail();
+        }
     }
 
     @Test
